+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - restore pane processes
 - user option for disabling pane process restoring
 - enable whitelisting processes that will be restored
+- expand readme with configuration options
 
 ### v0.0.4, 2014-08-26
 - restore pane layout for each window